==> python/data/1_JsonData.ipynb <==
# JSON 데이터 가공 시작
## 카테고리 정리

#### 100개가 넘는 카테고리가 존재하여 이를 통일된 규칙에 따라 정리한다.

### 필수 패키기 정리

import json
import pandas as pd

### 데이터 불러오기

# JSON 파일 경로
file_path = "/Users/isanghyeon/Desktop/json/cards_output_merged.json"

# 파일 열고 JSON 로딩
with open(file_path, 'r', encoding='utf-8') as f:
    data = json.load(f)

### 카테고리 구분

#### 카테고리 소카테고리 -> 중카테고리로 묶기

app = []

#모든가맹점
all_com = ["모든가맹점"]

#모빌리티
car_gasstation = ["주유소","충전소","주유"]
car_carAHipass = ["정비","차","중고차","하이패스","렌터카","자동차","차/중고차","자동차/하이패스"]

#대중교통
public_train = ["기차"]
public_texi = ["택시"]
public_expressbus = ["고속버스"]
public_transportation = ["대중교통"]
public_etc = ["교통"]

#통신
call_LG = ["LGU+"]
call_KT = ["KT"]
call_SKT = ["SKT"]
call_etc = ["통신"]

#생활
life_utilitybillsArental = ["공과금","렌탈","무이자할부","공과금/렌탈"]
life_studentAnursing = ["국민행복","문화센터","아이행복","어린이집","유치원","학습지","학원","생활"]
life_animal = ["동물병원","펫샵"]
life_etc = ["교육/육아"]

#쇼핑
shopping_conveniencestore = ["편의점"]
shopping_online = ["소셜커머스","홈쇼핑","온라인쇼핑"]
shopping_martAdepartmentstore = ["SPA브랜드","면세점","백화점","아울렛","인테리어","SSM","대형마트","전통시장"]
shopping_etc = ["마트/편의점","쇼핑"]

#외식/카페
foodAcafe_eatinghouse = ["일반음식점","주점","패밀리레스토랑","패스트푸드","점심","푸드"]
foodAcafe_delivery = ["배달앱"]
foodAcafe_cafe = ["카페"]
foodAcafe_dessert = ["베이커리","아이스크림"]
foodAcafe_etc = ["카페/디저트"]

#뷰티/피트니스
beautyAfitness_beauty = ['드럭스토어','헤어','화장품']
beautyAfitness_fitness = ['피트니스']
beautyAfitness_etc = ['뷰티/피트니스']

#금융/포인트
financeApoint_benefit = ["할인", "적립", "캐시백", "멤버십포인트", "해피포인트", "할인", "OK캐쉬백", "혜택 프로모션", "제휴/PLCC"]
financeApoint_finance = ["무실적", "지역", "증권사", "연회비지원", "금융", "CJ ONE", "카드사", "은행사", "BC TOP", "하이브리드"]
financeApoint_insurance = ["보험사","보험"]
financeApoint_pay = ["네이버페이", "삼성페이", "카카오페이", "PAYCO", "APP"]
financeApoint_charge = ["수수료우대", '해외이용', '해외직구']

#병원/약국
hospitalApharmacy_hospital = ['병원']
hospitalApharmacy_pharmacy = ['약국']
hospitalApharmacy_etc = ["병원/약국"]

#문화/취미
cultureAhobbies_OTT = ['디지털구독']
cultureAhobbies_movie = ['공연', '전시', '영화', '공연/전시']
cultureAhobbies_hobbies = ['음원사이트', '경기관람', '골프', '테마파크', '게임', '레저/스포츠', '도서']
cultureAhobbies_etc = ['OTT/영화/문화']

#숙박/항공
accommodationAflight_aviation = ['대한항공','아시아나항공','제주항공','에어부산','진에어','이스타항공','티웨이항공','저가항공','공항','항공권','항공마일리지']
accommodationAflight_lounge = ['PP','공항라운지','라운지키']
accommodationAflight_premium = ['프리미엄 서비스','바우처']
accommodationAflight_accommodationAflight = ['리조트','여행사','호텔','온라인 여행사',"여행/숙박"]



arr_key = ["모든가맹점"
           ,"주유소","자동차/하이패스"
           ,"기차","택시","고속버스","대중교통","교통"
           ,"LGU+","KT","SKT","통신"
           ,"공과금/렌탈","교육/육아","애완동물","교육/육아"
           ,'편의점','온라인쇼핑','마트/백화점','쇼핑'
           ,'음식점','배달','카페','디저트','외식/카페'
           ,'뷰티','피트니스','뷰티/피트니스'
           ,'혜택','금융','보험','간편결제','해외'
           ,'병원','약국','병원/약국'
           ,'OTT','영화','취미','OTT/영화/문화'
           ,'항공','공항라운지/PP','프리미엄',"여행/숙박"]

arr_value = [all_com
             ,car_gasstation,car_carAHipass
             ,public_train,public_texi,public_expressbus,public_transportation,public_etc
             ,call_LG,call_KT,call_SKT,call_etc
             ,life_utilitybillsArental,life_studentAnursing,life_animal,life_etc
             ,shopping_conveniencestore,shopping_online,shopping_martAdepartmentstore,shopping_etc
             ,foodAcafe_eatinghouse,foodAcafe_delivery,foodAcafe_cafe,foodAcafe_dessert,foodAcafe_etc
             ,beautyAfitness_beauty,beautyAfitness_fitness,beautyAfitness_etc
             ,financeApoint_benefit,financeApoint_finance,financeApoint_insurance,financeApoint_pay,financeApoint_charge
             ,hospitalApharmacy_hospital,hospitalApharmacy_pharmacy,hospitalApharmacy_etc
             ,cultureAhobbies_OTT,cultureAhobbies_movie,cultureAhobbies_hobbies,cultureAhobbies_etc
             ,accommodationAflight_aviation,accommodationAflight_lounge,accommodationAflight_premium,accommodationAflight_accommodationAflight]


### 데이터 가공 시작

# 카드 크롤링 데이터에서 ID값을 순차적으로 증가하는 값으로 변경
for idx, item in enumerate(data, start=1):
    item["id"] = idx
    
# 카테고리 소카테고리를 중카테고리로 key:value 형태로 변환
tag_dict = {}
for keywords, label in zip(arr_key, arr_value):
    tag_dict[keywords] = label
    
# 카테고리를 본격적으로 변환 시작
# 위에 선언한 카테고리에 없으면 app에 저장
app = []
for item in data:
    for benefit in item.get("benefits", []):
        original_place = benefit["place"]
        found = False
        
        for key, keywords in zip(arr_key, arr_value):
            if original_place in keywords:
                benefit["place"] = key
                found = True
                break

        if not found:
            if (original_place not in tag_dict and original_place not in app):
                app.append(original_place)

### JSON 형태로 저장

with open('/Users/isanghyeon/Desktop/json/test.json', 'w', encoding='utf-8') as f:
    json.dump(data, f, ensure_ascii=False, indent=2)
